+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v2Lib ChangeLog
 ***
-Version: `0.0.1-s20230312`
+Version: `0.0.2-s20230313`
 
 Changes:
 
-1. Fix root url redirect
+1. Allow external config